+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Add costume validation `valid->valid($closure, $message)` and costume filter `filter->filter($closure)`
+- Add method `Valid->raw($raw_rule)` to add raw valiation.
+- Add method __set to add new validation `Validator->__set` and `ValidPool->__set'.
 
 ## [0.6.2] - 2022-03-13
 ### Fixed